fixes of ControlOrderPage

+++ src/test/java/com/leroy/magportal/ui/pages/orders/ControlOrderPage.java
@@ -27,7 +27,7 @@ public class ControlOrderPage extends OrderCreatedPage {
     // Verifications
 
     @Step("Проверить, что кол-во 'Заказано' у {index}-ого товара равно {value}")
-    public ControlOrderPage shouldOrderedQuantityIs(int index, int value) throws Exception {
+    public ControlOrderPage shouldOrderedQuantityIs(int index, double value) throws Exception {
         index--;
         anAssert.isEquals(productCards.get(index).getOrderedQuantity(), String.valueOf(value),
                 "Неверное кол-во 'Заказано' у " + (index + 1) + "-ого товара");
@@ -35,7 +35,7 @@ public class ControlOrderPage extends OrderCreatedPage {
     }
 
     @Step("Проверить, что кол-во 'Собрано' у {index}-ого товара равно {value}")
-    public ControlOrderPage shouldPickedQuantityIs(int index, int value) throws Exception {
+    public ControlOrderPage shouldPickedQuantityIs(int index, double value) throws Exception {
         index--;
         anAssert.isEquals(productCards.get(index).getCollectedQuantity(),String.valueOf(value),
                 "Неверное кол-во 'Собрано' у " + (index + 1) + "-ого товара");
@@ -43,7 +43,7 @@ public class ControlOrderPage extends OrderCreatedPage {
     }
 
     @Step("Проверить, что кол-во 'Контроль' у {index}-ого товара равно {value}")
-    public ControlOrderPage shouldControlledQuantityIs(int index, int value) throws Exception {
+    public ControlOrderPage shouldControlledQuantityIs(int index, double value) throws Exception {
         index--;
         anAssert.isEquals(productCards.get(index).getControlledQuantity(), String.valueOf(value),
                 "Неверное кол-во 'Контроль' у " + (index + 1) + "-ого товара");
@@ -53,7 +53,7 @@ public class ControlOrderPage extends OrderCreatedPage {
 
     
     @Step("Проверить, кол-во 'Заказано', 'Собрано', 'Контроль' у {index}-ого товара равно {value}")
-    public ControlOrderPage orderPageVerifications(int index, int value) throws Exception {
+    public ControlOrderPage orderPageVerifications(int index, double value) throws Exception {
         index--;
         softAssert.isEquals(productCards.get(index).getOrderedQuantity(), String.valueOf(value),
                 "Неверное кол-во 'Заказано' у " + (index + 1) + "-ого товара");
